# Replace scratch buffers with pools