@mvdan 2022-03-20
Note that allocating buffers like `make([]byte, 8)` will end up on the stack unless they escape into the heap, by e.g. being returned or being fed to a generic `io.Reader.Read`. I imagine part of the slow-down here is caused by some of the lines you're replacing not being allocations in the first place.