@Stebalien 2022-03-20
Given that there are so few allocations anyways, I'm somewhat skeptical of this. Have you run any pprof profiles with `go tool pprof -alloc_objects` to verify?